+++ libs/serialization/test/src/unit_test_serialization_shared_ptr_polymorphic.cpp
@@ -57,7 +57,7 @@ public:
 
 	float GetValue() const override
 	{
-		return Base::GetValue() + b;
+		return Base::GetValue() + static_cast<float>(b);
 	}
 
 private:
